+++ src/main/java/com/blz/generic/TestMaximum.java
@@ -1,17 +1,17 @@
 package com.blz.generic;
 
 public class TestMaximum {
-    int x, y, z;
+    float x, y, z;
 
-    public TestMaximum(int x, int y, int z) {
+    public TestMaximum(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
     // testMaximum method created for checking max of three integers
-    public static Integer testMaximum(Integer x, Integer y, Integer z) {
-        Integer max = x;
+    public static Float testMaximum(Float x, Float y, Float z) {
+        Float max = x;
         if (y.compareTo(max) > 0)
             max = y;
         if (z.compareTo(max) > 0)
@@ -21,7 +21,7 @@ public class TestMaximum {
     }
 
     // printMax method created for displaying maximum values from the given 3 variables.
-    public static void printMax(Integer x, Integer y, Integer z, Integer max) {
+    public static void printMax(Float x, Float y, Float z, Float max) {
         System.out.printf("The max of %s, %s and %s is %s\n", x, y, z, max);
 
     }
